refactor: SetGoal refactor to match CA
